+++ src/features/funds/fundsAPI.js
@@ -2,12 +2,12 @@ import { apiSlice } from "../api/apiSlice";
 
 export const fundsApi = apiSlice.injectEndpoints({
   endpoints: (builder) => ({
-    getUserFundCategories: builder.query({
+    /* getUserFundCategories: builder.query({
       query: ({ email, page = 1, limit = 20, search = "" }) => {
         return `/funds/user-all-fund-category/lists?user=${ email }&page=${ page }&limit=${ limit }&search=${search}`
       },
       providesTags: ['Funds'],
-    }),
+    }), */
     getUserRecentFundsTransactions: builder.query({
       query: ({ email, page = 1, limit = 10, sort_by = '_id', sort_order = 'desc', search = "" }) => {
         return `/funds/user-funds?user=${ email }&page=${ page }&limit=${ limit }&sort_by=${ sort_by }&sort_order=${ sort_order }&search=${ search }`;
@@ -35,15 +35,8 @@ export const fundsApi = apiSlice.injectEndpoints({
       }),
       invalidatesTags: ['Funds', 'UserDetails'],
     }),
-    createUserFundCategory: builder.mutation({
-      query: (data) => ({
-        url: '/categories',
-        method: 'POST',
-        body: data
-      }),
-      invalidatesTags: ['Funds', 'UserDetails'],
-    }),
+    
   })
 })
 
-export const { useGetUserFundCategoriesQuery, useLazyGetUserFundCategoriesQuery, useGetUserRecentFundsTransactionsQuery, useGetUserCategoryFundListsQuery, useDeleteFundMutation, useAddFundMutation, useCreateUserFundCategoryMutation } = fundsApi;
+export const { useLazyGetUserFundCategoriesQuery, useGetUserRecentFundsTransactionsQuery, useGetUserCategoryFundListsQuery, useDeleteFundMutation, useAddFundMutation } = fundsApi;
